// src/interfaces/CustomerData.tsx
export interface CustomerData {
  fullName: string;
  email: string;
  document: string;
  zipCode: string;
  street: string;
  number: string;
  complement?: string;
  neighborhood: string;
  city: string;
  state: string;
  creditCardNumber: string;
  creditCardHolder: string;
  creditCardExpiration: string;
  creditCardSecurityCode: string;
}